Tweaked many prefabs

+++ Assets/RocketController.cs
@@ -65,7 +65,7 @@ public class RocketController : MonoBehaviour
     public float TimeThresholdForMinimalEvasion = 6;
 
     [Tooltip("Frames to stay in evasion mode after no longer being on a collision course with a friendly.")]
-    public int EvasionModeTime = 30;
+    public float EvasionModeTime = 30;
 
     [Tooltip("Distance in front of the rocket to start looking for friendlies on a collision cource - useful to avoid detecting itself.")]
     public float MinimumFriendlyDetectionDistance = 4;
